% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_data.r
\name{acf_data}
\alias{acf_data}
\title{Auto-correlation function (ACF)}
\usage{
acf_data(x, y = NULL, ...)
}
\arguments{
\item{y}{}
}
\value{
tibble
}
\description{
Computes data of ACF ou CCF
}
\examples{


x <- rnorm(100)

y <- rexp(100)

ccf_data(x,y)

}
