% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_num.r
\name{as_num}
\alias{as_num}
\title{Coerce character to numeric}
\usage{
as_num(x, thousand_mark = "\\\\.", decimal_mark = "\\\\,")
}
\arguments{
\item{x}{A character vector}

\item{thousand_mark}{A character single value to use as the thousand mark (default = '\ \.')}

\item{decimal_mark}{A character single value to use as the decimal mark (default = '\ \,')}
}
\value{
A numeric vector.
}
\description{
Coerce a character vector to numeric, by:
 \cr
 \cr  - Removing the thousand separator mark.
 \cr  - Replacing decimal mark with a dot.
 \cr  - Removing the character "$".
 \cr  - Removing the characters from the alphabet.
 \cr  - Trimming extra spaces.
 \cr  - Applying the function as.numeric.
}
\examples{

x <- "10.000,50"

as_num(x)

}
