% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_data.r
\name{cal_data}
\alias{cal_data}
\title{Data for a calendar plot}
\usage{
cal_data(dates, fills)
}
\arguments{
\item{dates}{vector of dates}

\item{fills}{numeric/character vector}
}
\value{
data.frame
}
\description{
Data for a calendar plot, from library ggcal
}
\examples{

date <- seq(as.Date("2017-02-01"), as.Date("2017-07-22"), by="1 day")

fill <- rnorm(length(date))

cal_data(date,fill)

}
