% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plot.r
\name{cal_plot}
\alias{cal_plot}
\title{Calendar plot}
\usage{
cal_plot(df)
}
\arguments{
\item{df}{data.frame from cal_data function}
}
\value{
ggplot2 plot
}
\description{
Calendar plot, from library ggcal
}
\examples{

date <- seq(as.Date("2017-02-01"), as.Date("2017-07-22"), by="1 day")

fill <- rnorm(length(mydate))

cal_plot(cal_data(date,fill))

}
