% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_acf.r
\name{calc_acf}
\alias{calc_acf}
\title{Auto-correlation function (ACF)}
\usage{
calc_acf(x, y = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
tibble
}
\description{
Computes data of ACF ou CCF
}
\examples{


x <- rnorm(100)

y <- rexp(100)

calc_acf(x,y)

}
