% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_association.r
\name{calc_association}
\alias{calc_association}
\title{Compute association coefficents}
\usage{
calc_association(
  x,
  y,
  type = c("chi-square", "contingency", "cramers-v", "fisher", "phi")
)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{type}{A character single value, that is the type of the association to be computed}
}
\value{
A numeric single value with the computed value.
}
\description{
Computation of association coefficients.
}
\details{
Currently, it is possible to compute the following metrics:

\cr - Contingency association coefficient
\cr - Chi-square test statistic
\cr - Cramer's V association coefficient
\cr - Fisher test statistic
\cr - Phi association coefficent
}
\examples{

calc_association(mtcars$vs,mtcars$am)

}
