% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc.r
\name{calc_auc}
\alias{calc_auc}
\title{Area under the curve computation}
\usage{
calc_auc(x, y, limits = NULL)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{limits}{A numeric vector of length 2 (default = range of 'x')}
}
\value{
A numeric single value with the computed AUC.
}
\description{
Computation of the area under the curve (AUC), by linear interpolation.
}
\examples{

x <- seq(-3,3,l = 100)

y <- dnorm(x)

plot(x,y)

calc_auc(x,y)

}
