% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_combination.r
\name{calc_combination}
\alias{calc_combination}
\title{Combination/Permutation computation}
\usage{
calc_combination(n, r, order_matter = FALSE, with_repetition = FALSE)
}
\arguments{
\item{n}{A numeric vector, that will be the total number of obervations}

\item{r}{A numeric vector, that will be the number of obervations to be selected}

\item{order_matter}{A boolean value, if TRUE the function will be a permutation, if not a combination (default = FALSE)}

\item{with_repetition}{A boolean value, if TRUE the function will consider that there will be repetitions (default = FALSE)}
}
\value{
A numeric single value.
}
\description{
Computation of number of combinations/permutations.
}
\examples{


calc_computation(6,2)

calc_computation(6,2,order_matter = TRUE,with_repetition = TRUE)

}
