% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_corr.r
\name{calc_corr}
\alias{calc_corr}
\title{Correlation computation}
\usage{
calc_corr(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A tibble (1x3) with the computed coefficients.
}
\description{
Computation of the Pearson, Kendall and Spearman correlations coefficients.
}
\examples{

x <- rnorm(100)

y <- rnorm(100)

plot(x,y)

calc_corr(x,y)


}
