% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_correlation.r
\name{calc_correlation}
\alias{calc_correlation}
\title{Correlation/association computation}
\usage{
calc_correlation(x, y, type = "pearson")
}
\arguments{
\item{x}{A numeric vector}

\item{type}{A character single value, that is the type of the correlation to be computed}
}
\value{
A numeric single value with the computed value.
}
\description{
Computation of correlation/association coefficients.
}
\details{
Currently, it is possible to compute the following metrics:
}
\examples{

x <- rnorm(100,1)

calc_correlation(x)

}
