% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cv.r
\name{calc_cv}
\alias{calc_cv}
\title{Coefficient of variation computation}
\usage{
calc_cv(x, as_perc = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{as_perc}{A boolean value, if TRUE the function will multiply the CV by 100 (default = FALSE)}
}
\value{
A numeric single value with the computed CV.
}
\description{
Computation of the coefficient of variation (CV).
}
\examples{

x <- rnorm(100,1)

calc_cv(x)

}
