% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_date_aux.r
\name{calc_date_aux}
\alias{calc_date_aux}
\title{Auxiliar date variables computation}
\usage{
calc_date_aux(df, dt_var)
}
\arguments{
\item{df}{A single data.frame}

\item{dt_var}{A date variable from a data.frame}
}
\value{
A tibble with the original data and the new computed variables.
}
\description{
Computation of auxiliar date variables:
 \cr
 \cr - mon_abb: a factor with the abbreviated name of the month;
 \cr - mon_lbl: a factor with the complete name of the month;
 \cr - mon_num: the number of the month;
 \cr - day_num: the number of the day within the month;
 \cr - year_num: the number of the year;
 \cr - year_lbl: a factor with the number of the year;
 \cr - week_num: the number of complete seven day periods that have occurred between the date and January 1st, plus one;
 \cr - epiweek_num: the US CDC version of epidemiological week. Starts on Sunday;
 \cr - isoweek_num:  the week as it would appear in the ISO 8601 system, which uses a reoccurring leap week. Starts on Monday;
 \cr - wday_abb: a factor with the week day abbreviated name;
 \cr - wday_lbl: a factor with the week day complete name.
}
\examples{

df <- data.frame(dt = lubridate::dmy("01/01/01"))

calc_date_aux(df,dt)

}
