% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_date_diff.r
\name{calc_date_diff}
\alias{calc_date_diff}
\title{Difference between dates}
\usage{
calc_date_diff(date1, date2, add = 0, unit = "days")
}
\arguments{
\item{date1}{date number 1}

\item{date2}{date number 2}

\item{add}{auxiliar number to sum to the difference (date2 - date1 + add)}

\item{unit}{unit of the date difference}
}
\value{
numeric value
}
\description{
Difference between dates (date2 - date1)
}
\examples{

date1 <- lubridate::dmy("01/05/1998")

date2 <- lubridate::dmy("21/11/2018")

calc_date_diff(date1 = date1,date2 = date2,units = "days")
}
