% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.r
\name{calc_error}
\alias{calc_error}
\title{Compute error metrics}
\usage{
calc_error(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}
}
\value{
A tibble with the computed error metrics.
}
\description{
Compute error metrics
}
\examples{

x <- rnorm(100)

y <- rnorm(100)

plot(x,y)

calc_error(x,y)

}
