% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kurtosis.r
\name{calc_kurtosis}
\alias{calc_kurtosis}
\title{Kurtosis computation}
\usage{
calc_kurtosis(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric single value with the computed value.
}
\description{
Computation of the kurtosis, a measure of the "peakedness".
 This metric can be used to determine the shape of a dataset,
  and whether it is platykurtic (flatter than a normal distribution),
   mesokurtic (same kurtosis as a normal distribution),
    or leptokurtic (more peaked than a normal distribution).
}
\examples{

x <- rnorm(100,1)

calc_kurtosis(x)

}
