% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.r
\name{calc_mean}
\alias{calc_mean}
\title{Mean}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
data.frame
}
\description{
Compute arithmetic, geometric and harmonic mean
}
\examples{

x <- rnorm(100)

calc_mean(x)

}
