% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_modality.r
\name{calc_modality}
\alias{calc_modality}
\title{Modality computation}
\usage{
calc_modality(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A single value.
}
\description{
The number of modes a dataset has.
It can be used to identify whether a dataset is unimodal, bimodal, or multimodal.
}
\examples{

table(mtcars$cyl)

calc_mode(mtcars$cyl)

calc_modality(mtcars$cyl)

}
