% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.r
\name{calc_mode}
\alias{calc_mode}
\title{Mode for character values}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
character value
}
\description{
Find the mode
}
\examples{

table(mtcars$cyl)

calc_mode(mtcars$cyl)

}
