% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_peak_density.r
\name{calc_peak_density}
\alias{calc_peak_density}
\title{Mode for numeric values}
\usage{
calc_peak_density(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
numeric value
}
\description{
Find the high density value
}
\examples{

x <- rnorm(100)

calc_peak_density(x)

}
