% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perc.r
\name{calc_perc}
\alias{calc_perc}
\title{Compute a percentage}
\usage{
calc_perc(df, grp_var, main_var = NULL)
}
\arguments{
\item{df}{A single data.frame}

\item{grp_var}{The grouping variable(s) based on which the percentages will be computed}

\item{main_var}{An optional variable, when provided, will be utilized to compute the percentage for each of its own levels}
}
\value{
A data.frame summarizing the counts and percentages within groups. The output is arranged in descending order of counts within each group.
}
\description{
Compute  the percentage based on one or multiple variables.
}
\examples{

#without main_var
calc_perc(mtcars,grp_var = c(cyl,vs))

#main_var within grp_var
calc_perc(mtcars,grp_var = c(cyl,vs),main_var = vs)

#main_var not within grp_var
calc_perc(mtcars,grp_var = c(cyl),main_var = vs)
}
