% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_skewness.r
\name{calc_skewness}
\alias{calc_skewness}
\title{Skewness computation}
\usage{
calc_skewness(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric single value with the computed value.
}
\description{
Computation of the skewness, a measure of the asymmetry.
This metric can be used to determine the shape of a variable,
 and whether it is symmetric or skewed.
}
\examples{

x <- rnorm(100,1)

calc_skewness(x)

}
