% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_mode.R
\name{cat_mode}
\alias{cat_mode}
\title{Mode for character values}
\usage{
cat_mode(x)
}
\arguments{
\item{x}{character vector}
}
\value{
character value
}
\description{
Find the mode
}
\examples{

table(mtcars$cyl)

cat_mode(mtcars$cyl)

}
