% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_summary.r
\name{cat_summary}
\alias{cat_summary}
\title{Categoric summary}
\usage{
cat_summary(x)
}
\arguments{
\item{x}{characther/factor vector}
}
\value{
tibble
}
\description{
Compute summary statistics for categoric variable
}
\examples{

x <- c(sample(letters,100,re = T),NA)

x

cat_summary(x)

}
