% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccf_data.r
\name{ccf_data}
\alias{ccf_data}
\title{Cross-correlation function (CCF)}
\usage{
ccf_data(x, y, ...)
}
\value{
tibble
}
\description{
Computes data of CCF
}
\examples{


x <- rnorm(100)

y <- rexp(100)

ccf_data(x,y)

}
