% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count2.r
\name{count2}
\alias{count2}
\title{Compute the frequency of counts, giving a group variable}
\usage{
count2(df, grp_var)
}
\arguments{
\item{df}{A single data.frame}

\item{grp_var}{A  variable from a data.frame}
}
\value{
A data.frame
}
\description{
It counts the frequency of the counts, gibing a initial frequency count of the number of occurrences of each group size.
}
\examples{

count2(mtcars,cyl)

}
