% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.r
\name{count_na}
\alias{count_na}
\title{Count the number of NA observations}
\usage{
count_na(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A numeric value
}
\description{
Count the number of NA observations
}
\examples{

count_na(c(2,2,2))

count_na(c(2,2,NA))

}
