% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_vars.R
\name{date_vars}
\alias{date_vars}
\title{Criação de variáveis derivadas de data}
\usage{
date_vars(df, date_var)
}
\arguments{
\item{df}{data.frame}

\item{date_var}{variável de data}
}
\value{
data.frame
}
\description{
Cria variáveis a partir de uma data, como: ano, mês e dia.
}
\examples{

date_vars(df,dt_saida)

}
