% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_summary.r
\name{df_summary}
\alias{df_summary}
\title{Data summary}
\usage{
df_summary(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
tibble
}
\description{
Compute summary statistics for data
}
\examples{

mtcars

df_summary(mtcars)

}
