% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_date.R
\name{diff_date}
\alias{diff_date}
\title{Difference between dates}
\usage{
diff_date(first_date, last_date)
}
\arguments{
\item{last_date}{}
}
\value{
numeric value
}
\description{
Difference between dates
}
\examples{



}
