% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dttm_diff.r
\name{dttm_diff}
\alias{dttm_diff}
\title{Difference between two dates}
\usage{
dttm_diff(
  date1,
  date2,
  add = 0,
  unit = c("days", "secs", "mins", "hours", "weeks")
)
}
\arguments{
\item{date1}{A date single value}

\item{date2}{A date single value}

\item{add}{A numeric single value (default = 0)}

\item{unit}{A character single value with the unit of the date result (secs, mins, hours, days, weeks)}
}
\value{
A numeric single value with the computed difference between the two dates.
}
\description{
Computation of the difference between two dates (date2 - date1 + add).
}
\examples{

date1 <- lubridate::dmy("01/05/1998")

date2 <- lubridate::dmy("21/11/2018")

dttm_diff(date1 = date1,date2 = date2,unit = "days")
}
