% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_plot.r
\name{dual_plot}
\alias{dual_plot}
\title{Dual axis plot}
\usage{
dual_plot(
  df,
  x_axis,
  y_left,
  y_right,
  y_breaks = NULL,
  col_left = "forestgreen",
  col_right = "chocolate2",
  legend_pos = "top",
  ...
)
}
\arguments{
\item{df}{data.frame}

\item{x_axis}{name of the x axis variable}

\item{y_left}{name of the y axis variable for the left side}

\item{y_right}{name of the y axis variable for the right side}

\item{y_breaks}{numeric vector of the y axis left breaks}

\item{col_left}{name of the color for the left side}

\item{col_right}{name of the color for the right side}

\item{legend_pos}{legend position}
}
\value{
ggplot2 object
}
\description{
Rescale data with linear scaling, non-linear relationship between variables may cause wrong interpretations
}
\examples{

x <- seq(0,1,l = 100)

y <- cumsum(rnorm(100))

z <- y^3 + rnorm(100)

data <- tibble(x,y,z)

dual_plot(data,"x","y","z")

}
