% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.r
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{expand_grid without repetition}
\usage{
expand_grid_unique(x, y, include_equals = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{include_equals}{logical indicator to include equal pairs or not}
}
\value{

}
\description{

}
\examples{

expand_grid_unique(x = 1:6,y = 1:8)

}
