% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_flip_y_title.r
\docType{data}
\name{flip_y_title}
\alias{flip_y_title}
\title{Flip y axis title}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
flip_y_title
}
\value{
theme object
}
\description{
axis.title.y rotated
}
\examples{
library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + flip_y_title

}
\keyword{datasets}
