% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.r
\name{format_num}
\alias{format_num}
\title{Format numeric data}
\usage{
format_num(
  x,
  digits = 2,
  decimal_mark = ".",
  thousand_mark = ",",
  br_mark = FALSE
)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of decimals digits}

\item{decimal_mark}{decimal mark}

\item{thousand_mark}{thousand mark}

\item{br_mark}{if TRUE, decimal_mark = "," and thousand_mark = "." (default = FALSE)}
}
\value{
formatted data
}
\description{
Format data as "10.500,25"
}
\examples{

format_num(12345.67)

}
