% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.r
\name{format_p_value}
\alias{format_p_value}
\title{Format p value}
\usage{
format_p_value(p_value, p_value_min = NULL, ...)
}
\arguments{
\item{p_value}{A numeric vector}

\item{p_value_min}{A numeric single value (default = NULL), that will be the lower bound}
}
\value{
formatted data
}
\description{
Format p value
}
\examples{

format_p_value(c(.001,.00000001),p_value_min = 0.001)

}
