% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_scale.r
\name{format_scale}
\alias{format_scale}
\title{Scaling data}
\usage{
format_scale(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{numeric vector}

\item{new_min}{numeric value of the inferior limit (0 by default)}

\item{new_max}{numeric value of the superior limit (1 by default)}
}
\value{
numeric vector
}
\description{
Rescale data
}
\examples{

x <- seq(-3,3,l = 100)

range(x)

y <- format_scale(x)

range(y)

plot(x,y)

}
