% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_mean.r
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean}
\usage{
geometric_mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Geometric mean
}
\description{
Calculates geometric mean
}
\examples{

x <- rexp(100,.5)

geometric_mean(x)

}
