% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_mean.r
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Harmonic mean}
\usage{
harmonic_mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Harmonic mean
}
\description{
Calculates harmonic mean
}
\examples{

x <- rexp(100,.5)
harmonic_mean(x)

}
