% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank.r
\name{is_blank}
\alias{is_blank}
\title{Conditional check for a blank value}
\usage{
is_blank(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is blank.
}
\examples{

is_blank("")

is_blank(NA_character_)

is_blank(" ")

is_blank("-")

}
