% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.r
\name{is_date}
\alias{is_date}
\title{Conditional check for a date variable}
\usage{
is_date(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the vector is a date.
}
\examples{

is_date(Sys.time())

is_date(Sys.Date())

}
