% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_datetime.r
\name{is_datetime}
\alias{is_datetime}
\title{Datetime variable}
\usage{
is_datetime(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
Boolean vector
}
\description{
return TRUE if the vector is a datetime
}
\examples{


}
