% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.r
\name{is_even}
\alias{is_even}
\title{Even number}
\usage{
is_even(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector
}
\description{
return TRUE if the number is even
}
\examples{

is_even(4)

}
