% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.r
\name{is_integer}
\alias{is_integer}
\title{Conditional check for an integer value}
\usage{
is_integer(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is a integer,
 not considering their variable type but the value itself.
}
\examples{

is_integer(1)

is_integer(1.1)

}
