% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_negative.r
\name{is_negative}
\alias{is_negative}
\title{Conditional check for a negative value}
\usage{
is_negative(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is negative.
}
\examples{

is_negative(4)

is_negative(-4)

}
