% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.r
\name{is_odd}
\alias{is_odd}
\title{Conditional check for an odd value}
\usage{
is_odd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is odd.
}
\examples{

is_odd(4)

is_odd(3)

}
