% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.r
\name{is_outlier}
\alias{is_outlier}
\title{Outlier}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector
}
\description{
return TRUE if the value is an outlier (boxplot criteria)
}
\examples{

x <- c(rnorm(10),10)

x

is_outlier(x)

}
