% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_positive.r
\name{is_positive}
\alias{is_positive}
\title{Conditional check for a positive value}
\usage{
is_positive(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is positive.
}
\examples{

is_positive(4)

is_positive(-4)

}
