% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_string.r
\name{is_string}
\alias{is_string}
\title{Conditional check for an string value}
\usage{
is_string(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is a string (character or factor).
}
\examples{

is_string("A")

}
