% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_weekend.r
\name{is_weekend}
\alias{is_weekend}
\title{Conditional check if a date is a weekend}
\usage{
is_weekend(x)
}
\arguments{
\item{x}{A date vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the date vector is a weekend (saturday/sunday).
}
\examples{

is_weekend(x = Sys.time())

is_weekend(x = Sys.Date())

}
