% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_zero.r
\name{is_zero}
\alias{is_zero}
\title{Conditional check for a value equal to zero}
\usage{
is_zero(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is equal to zero.
}
\examples{

is_zero(4)

is_zero(0)

}
