% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnot_na.r
\name{isnot_na}
\alias{isnot_na}
\title{Opposite of is.na}
\usage{
isnot_na(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
Boolean vector
}
\description{
Check if the value is not a NA
}
\examples{

isnot_na(2)

isnot_na(NA)

}
