% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_seq.r
\name{mult_seq}
\alias{mult_seq}
\title{Sequence by a multiple}
\usage{
mult_seq(x, m = 1)
}
\arguments{
\item{x}{numeric vector.}

\item{m}{multiple.}
}
\value{
sequence
}
\description{
Gets sequence by a multiple, with the minimum and maximum of data
}
\examples{
x <- rnorm(100)

mult_seq(x,m = .5)

}
