% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_spread.r
\name{mult_spread}
\alias{mult_spread}
\title{Spread with multiple variables}
\usage{
mult_spread(df, key, value)
}
\arguments{
\item{df}{data.frame}

\item{key}{key variable name}

\item{value}{value variable name}
}
\value{
tibble
}
\description{
Compute summary statistics for data
}
\examples{


}
