% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_bg.R
\docType{data}
\name{no_bg}
\alias{no_bg}
\title{Transparent background}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
no_bg
}
\value{
theme object
}
\description{
Transparent background for ggplot2
}
\examples{
library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + no_bg

}
\keyword{datasets}
