% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Opposite of %in%}
\usage{
not_in(x, y)
}
\arguments{
\item{x}{single value}

\item{y}{vector}
}
\value{
boolean
}
\description{
Check if a x are not in y
}
\examples{

not_in(a, letters)

}
