% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_na.R
\name{not_na}
\alias{not_na}
\title{Opposite of is.na}
\usage{
not_na(x)
}
\arguments{
\item{x}{single value or vector}
}
\value{
boolean
}
\description{
Check if the value is not a NA
}
\examples{

not_na(2)

not_na(NA)

}
