% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_convert.r
\name{num_convert}
\alias{num_convert}
\title{Numeric vector}
\usage{
num_convert(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Numeric vector
}
\description{
Coerce a vector to numeric: removing thousand separator, replace comma with dots and applytins as.numeric()
}
\examples{

x <- "10.000,50"

num_convert(x)

}
