% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_corr.r
\name{num_corr}
\alias{num_corr}
\title{Numeric correlation}
\usage{
num_corr(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
Numeric vector
}
\description{
Compute pearson, kendall and spearman correlation coefficient
}
\examples{

x <- rnorm(100)

y <- rnorm(100)

plot(x,y)

num_corr(x,y)


}
