% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_format.r
\name{num_format}
\alias{num_format}
\title{Format numeric data}
\usage{
num_format(x, digits = 2)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of decimals digits}
}
\value{
formatted data
}
\description{
Format data as "10.500,25"
}
\examples{

num_format(12345.67)

}
