% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_mean.r
\name{num_mean}
\alias{num_mean}
\title{Mean}
\usage{
num_mean(x)
}
\arguments{
\item{x}{numeric vector
#'}
}
\value{
Numeric vector
}
\description{
Compute arithmetic, geometric and harmonic mean
}
\examples{

x <- rnorm(100)

num_mean(x)

}
