% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_mode.r
\name{num_mode}
\alias{num_mode}
\title{Mode for numeric values}
\usage{
num_mode(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric value
}
\description{
Find the high density value
}
\examples{

x <- rnorm(100)

num_mode(x)

}
