% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_summary.r
\name{num_summary}
\alias{num_summary}
\title{Numeric summary}
\usage{
num_summary(x, minimal = F)
}
\arguments{
\item{x}{numeric vector}

\item{minimal}{minimal version (default = F)}
}
\value{
tibble
}
\description{
Compute summary statistics for numeric variable
}
\examples{

x <- c(rnorm(10),NA,10)

x

num_summary(x)

}
