% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_to_string.R
\name{obj_to_string}
\alias{obj_to_string}
\title{Object to string}
\usage{
obj_to_string(x)
}
\arguments{
\item{x}{}
}
\value{
vector
}
\description{
Get the name of the object and return as a string
}
\examples{


x <- rnorm(100)

obj_to_string(x)


}
