% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_div.R
\name{pal_div}
\alias{pal_div}
\title{Palettes for divergent scale}
\usage{
pal_div(name = "young_sheldon", reverse = FALSE, show_palettes = FALSE)
}
\arguments{
\item{name}{A character single value, that is the pallete name}

\item{reverse}{A boolean value, if TRUE the function will reverse the pallete order (default = FALSE)}

\item{show_palettes}{A boolean value, if TRUE the function will export the names of all the palettes (default = FALSE)}
}
\value{
A character vector with the colors hex codes.
}
\description{
Palette for a divergent scale, with the following palettes:
 \cr
 \cr - arcane: 5 colors, inspired by the animation Arcane.
 \cr - auto_compadecida: 6 colors, inspired by the movie O Auto da Compadecida.
 \cr - blue_eye_samurai: 6 colors, inpired by the TV series Blue Eye Samurai.
 \cr - casa_de_papel: 7 colors, inspired by the TV series La Casa de Papel.
 \cr - deadly_class: 6 colors, inspired by the TV series Deadly Class.
 \cr - dexter: 6 colors, inspired by the TV series Dexter.
 \cr - godzilla_vs_kong: 9 colors, inspired by the movie Godzilla vs Kong.
 \cr - good_place: X colors, inspired by the TV series The Good Place.
 \cr - homeland: 9 colors, inspired by the TV series Homeland.
 \cr - laura_dean: 8 colors, inspired by the graphic novel Laura Dean Keeps Breaking Up with Me.
 \cr - racionais: 7 colors, inspired by the CD cover
 'Nada Como Um Dia Após o Outro Dia' of the brazilian rap group Racionais MC's.
 \cr - sandman: 5 colors, inspired by the graphic novel Sandman.
 \cr - svott_pilgrim: 6 colors, inspired by the movie Scott Pilgrim vs the World.
 \cr - the_americans: 7 colors, inspired by the TV series The Americans.
 \cr - the_expanse: 8 colors, inspired by the TV series The Expanse.
 \cr - the_marvels: 7 colors, inspired by the movie The Marvels.
 \cr - vox_machina: 7 colors, inspired by the animation The Legend of Vox Machina.
 \cr - wanda_vision: 9 colors, inspired by the TV series WandaVision.
 \cr - white_lotus: 8 colors, inspired by the TV series The White Lotus.
}
\examples{

scales::show_col(pal_div("sandman"))

}
