% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_two.R
\name{pal_two}
\alias{pal_two}
\title{Palettes with two constrast colors}
\usage{
pal_two(
  name = c("barbie", "boba_fett", "blade_runner_2049", "dear_white_people",
    "emily_in_paris", "good_omens", "hightown", "jaws", "notorius", "saint_maud",
    "shirley", "spirited", "the_english", "the_expanse", "the_gilded_age", "the_help",
    "titans", "watchmen", "winning_time"),
  reverse = FALSE
)
}
\arguments{
\item{name}{A character single value, that is the pallete name}

\item{reverse}{A boolean value, if TRUE the function will reverse the pallete order (default = FALSE)}
}
\value{
A character vector with the colors hex codes.
}
\description{
Palette with two constrast colors, with the following palettes:
 \cr
 \cr - barbie: Inspired by the movie Barbie.
 \cr - boba_fett: Inspired by the TV series Boba Fett.
 \cr - blade_runner_2049: Inspired by the movie Blade Runner 2049.
 \cr - dear_white_people: Inspired by the TV series Dear White People.
 \cr - emily_in_paris: Inspired by the TV series Emily in Paris.
 \cr - good_omens: Inspired by the TV series Good Omens.
 \cr - hightown: Inspired by the TV series Hightown.
 \cr - jaws: Inspired by the movie Jaws.
 \cr - notorious: Inspired by the movie Notorious.
 \cr - saint_maud: Inspired by the movie Saint Maud.
 \cr - shirley: Inspired by the movie Shirley.
 \cr - spirited: Inspired by the movie Spirited.
 \cr - the_english: Inspired by the movie The English.
 \cr - the_expanse: Inspired by the TV series The Expanse.
 \cr - the_gilded_age: Inspired by the TV series The Gilded Age.
 \cr - the_help: Inspired by the movie The Help.
 \cr - titans: Inspired by the TV series Titans.
 \cr - watchmen: Inspired by the TV series Watchmen.
 \cr - winning_time: Inspired by the TV series Winning Time.
}
\examples{

scales::show_col(pal_two("boba_fett"))

}
