% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_example.R
\name{palette_example}
\alias{palette_example}
\title{Makes a visual representation of a palette}
\usage{
palette_example(palette)
}
\arguments{
\item{palette}{A character vector, that}
}
\value{
A ggplot2 plot.
}
\description{
Makes a visual representation of a palette.
}
\examples{

palette_example(palette = palette_div(name = "arcane"))

}
