% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_seq.R
\name{palette_seq}
\alias{palette_seq}
\title{Palettes for sequential scale}
\usage{
palette_seq(
  name = c("andor", "breaking_bad", "cyberpunk", "raised_by_wolves", "sucession"),
  reverse = FALSE
)
}
\arguments{
\item{name}{A character vector, that is the pallete name}

\item{reverse}{A boolean value, if TRUE the function will reverse the pallete order (default = FALSE)}
}
\value{
A character vector with the colors hex codes.
}
\description{
Palettes for a sequential scale, with the following palettes:
 \cr
 \cr - andor: 6 colors, inspired by the TV series Andor (Star Wars).
 \cr - breaking_bad: 5 colors, inspired by the TV series Breaking Bad.
 \cr - cyberpunk: 5 colors, inspired by the videogame Cyberpunk 2077.
 \cr - raised_by_wolves: 6 colors, inspired by the TV series Raised by Wolves.
 \cr - sucession: 6 colors, inspired by the TV series Sucession.
}
\examples{

palette_example(palette = palette_seq(name = "cyberpunk"))

}
