% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_two.R
\docType{data}
\name{pals_two}
\alias{pals_two}
\title{List of binary colors palettes}
\format{
An object of class \code{list} of length 24.
}
\usage{
pals_two
}
\description{
List of binary colors palettes
}
\keyword{datasets}
