% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_text.R
\name{parse_text}
\alias{parse_text}
\title{Parse text}
\usage{
parse_text(x)
}
\arguments{
\item{x}{char vector}
}
\value{
prints a image
}
\description{
Extract only text from a string
}
\examples{

parse_text("1ABCF45Z89")


}
