% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_flip_y_title.r
\docType{data}
\name{plt_flip_y_title}
\alias{plt_flip_y_title}
\title{ggplot2: Flip y axis title}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
plt_flip_y_title
}
\value{
A ggplot2 object.
}
\description{
The y axis title with a horizontal rotation.
}
\examples{
library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_flip_y_title

}
\keyword{datasets}
