% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_identity_line.R
\name{plt_identity_line}
\alias{plt_identity_line}
\title{Add a identity line to ggplot2 scatter plot}
\usage{
plt_identity_line(color = "black", linetype = "dashed")
}
\arguments{
\item{color}{line color (default = 'black')}

\item{linetype}{line type (default = 'dashed')}
}
\value{
ggplot2 theme
}
\description{
Add a identity line to ggplot2 scatter plot
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(qsec,mpg))+
  geom_point()

plot

plot + plt_identity_line()

}
