% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_no_background.r
\docType{data}
\name{plt_no_background}
\alias{plt_no_background}
\title{Transparent background}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
plt_no_background
}
\value{
theme object
}
\description{
Transparent background for ggplot2
}
\examples{
library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_no_background

}
\keyword{datasets}
