% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_no_grid.r
\docType{data}
\name{plt_no_grid}
\alias{plt_no_grid}
\title{ggplot2: Removal of grid}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 1.
}
\usage{
plt_no_grid
}
\value{
A ggplot2 object.
}
\description{
Removal of the grid from a ggplot2 plot.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_no_grid

}
\keyword{datasets}
