% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_no_text.r
\docType{data}
\name{plt_no_text}
\alias{plt_no_text}
\title{ggplot2: Removal of all text and ticks from both axis}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 2.
}
\usage{
plt_no_text
}
\value{
A ggplot2 object.
}
\description{
Removal of all text and ticks from both axis, in a ggplot2 plot.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_no_text

}
\keyword{datasets}
