% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_pinpoint.R
\name{plt_pinpoint}
\alias{plt_pinpoint}
\title{ggplot2: Add a single point with two lines in a ggplot2}
\usage{
plt_pinpoint(
  x,
  y,
  color = "firebrick3",
  size = 3,
  linetype = "dashed",
  linewidth = 0.75
)
}
\arguments{
\item{x}{A numeric single value}

\item{y}{A numeric single value (default = 3)}

\item{color}{A character single value (default = 'firebrick3'), that will be the line color}

\item{linetype}{A character single value (default = 'dashed'), that will be the line type}

\item{linewidth}{A numeric single value (default = 0.75), that will be the line width}
}
\value{
A ggplot2 object.
}
\description{
Add a single point in a ggplot2 scatter plot.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(qsec,mpg))+
  geom_point()

plot

plot + plt_pinpoint(x = 18,y = 25)

}
