% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_regression_line.R
\name{plt_regression_line}
\alias{plt_regression_line}
\title{ggplot2: Add a regression line to a scatter plot}
\usage{
plt_regression_line(color = "black", linetype = "dashed", linewidth = 0.75)
}
\arguments{
\item{color}{A character single value (default = 'black'), that will be the line color}

\item{linetype}{A character single value (default = 'dashed'), that will be the line type}

\item{linewidth}{A numeric single value (default = 0.75), that will be the line width}
}
\value{
A ggplot2 object.
}
\description{
Add a regression line to a ggplot2 scatter plot.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(qsec,mpg))+
  geom_point()

plot

plot + plt_regression_line()

}
