% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_scale_auto.r
\name{plt_scale_auto}
\alias{plt_scale_auto}
\title{Auto scale}
\usage{
plt_scale_auto(
  axis = "x",
  n_ticks = 10,
  expand = 0.01,
  thousand_mark = ".",
  decimal_mark = ","
)
}
\arguments{
\item{axis}{axis indicator: x (default), y or xy}

\item{n_ticks}{number of ticks}

\item{expand}{expanded space}
}
\value{
ggplot2 object
}
\description{
Provide automatic scaling to ggplot2
}
\examples{
library(ggplot2)

plot <-
ggplot(mtcars,aes(hp,drat))+
geom_point()

plot+
 plt_scale_auto("x",n_ticks = 10)+
 plt_scale_auto("y",n_ticks = 10)


}
