% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_scale_y_mirror.R
\name{plt_scale_y_mirror}
\alias{plt_scale_y_mirror}
\title{ggplot2: Mirror the y axis on the right side}
\usage{
plt_scale_y_mirror(breaks, labels = NULL, expand = c(0.01, 0), ...)
}
\arguments{
\item{breaks}{A  vector, that will be the breaks in the y axis}

\item{labels}{A  vector (default = NULL), that will be the labels in the axis, if not provided 'breaks' will be used instead}

\item{expand}{A numeric vector (default = c(0.01,0)), that is the range expansion used to add some space between the data and the plot border}
}
\value{
A ggplot2 object.
}
\description{
Mirror the ticks and text from the y axis on the right side.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_scale_y_mirror(breaks = seq(0,40,4))

}
