% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_theme_x.r
\name{plt_theme_x}
\alias{plt_theme_x}
\title{ggplot2 theme}
\usage{
plt_theme_x(base_size = 14, base_family = "", margin = 0.35)
}
\value{
ggplot2 theme
}
\description{
theme_bw with no minor grid and major y grid
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_theme_x()

}
