% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_theme_xy.r
\name{plt_theme_xy}
\alias{plt_theme_xy}
\title{ggplot2: A black and white theme without minor grid lines}
\usage{
plt_theme_xy(base_size = 14, base_family = "", margin = 0.35)
}
\arguments{
\item{base_size}{A numeric single value (default = 14), that will be the base font size}

\item{base_family}{A character single value (default = ''), that will be the base font family}

\item{margin}{A numeric single value (default = 0.35), that will be the plot margin}
}
\value{
A ggplot2 object.
}
\description{
A theme with no minor grids in both axes
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_theme_xy()

}
