% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_theme_y.r
\name{plt_theme_y}
\alias{plt_theme_y}
\title{ggplot2 theme}
\usage{
plt_theme_y(base_size = 14, base_family = "", margin = 0.35)
}
\arguments{
\item{base_size}{A numeric single value (default = 14), that will be the base font size}

\item{margin}{A numeric single value (default = 0.35), that will be the plot margin}
}
\value{
ggplot2 theme
}
\description{
theme_bw with no minor grid and major x grid
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_theme_y()

}
