% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_water_mark.r
\name{plt_water_mark}
\alias{plt_water_mark}
\title{Water mark}
\usage{
plt_water_mark(png_file = NULL, png_path = NULL)
}
\arguments{
\item{png_file}{png file, already imported in R}

\item{png_path}{png path to import the png file}
}
\value{
object to use in ggplot2
}
\description{
Import png file as watermark for ggplot2
}
\examples{
#'
library(ggplot2)

#A local png file is necessary

ggplot(mtcars)+
  plt_water_mark(file_path = path_to_local_png)
}
