% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_date.r
\name{range_date}
\alias{range_date}
\title{Date range}
\usage{
range_date(x)
}
\arguments{
\item{x}{date vector.}
}
\value{
date range
}
\description{
Gets range of a date vector
}
\examples{
x <- seq(as.Date("1910/1/1"), as.Date("1911/1/1"), "days")

range_date(x)

}
