% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.r
\name{replace_na}
\alias{replace_na}
\title{Replace a NA value}
\usage{
replace_na(x, na_value)
}
\arguments{
\item{x}{A  vector}

\item{na_nalue}{A  single value, that will replace NA values}
}
\value{
A numeric value
}
\description{
Replace a NA value
}
\examples{

replace_na(c(2,2,NA,NA,NA),1)

}
