% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_unique.r
\name{row_number_unique}
\alias{row_number_unique}
\title{Create a rank given unique values from a variable}
\usage{
row_number_unique(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A numeric vector with rank, given the unique values from argument 'x'.
}
\description{
Get the row number for unique values.
}
\examples{

row_number_unique(mtcars$gear)

}
