% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpearson.r
\name{rpearson}
\alias{rpearson}
\title{Linear correlated data generator}
\usage{
rpearson(n = 25, pearson = 0.5, tol = 0.1, mean = 0, sd = 1)
}
\arguments{
\item{n}{A numeric integer single value (default = 25), that will be the number of observations to simulate}

\item{pearson}{A numeric single value (default = 0.50), that will be the expected Pearson linear correlation coefficient}

\item{tol}{A numeric single value (default = 0.10), that will be the tolerance between the simulated and estimated correlation}

\item{mean}{A numeric single value (default = 0), that will be the mean of the simulated data}

\item{sd}{A numeric single value (default = 1), that will be the standard deviation of the simulated data}
}
\value{
Data.frame with 2 numeric correlated series (x and y)
}
\description{
Simulates linear correlated data
}
\examples{

df <- rpearson(n = 100, pearson = .8, mean = 3)

head(df)

plot(df)

}
