% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale01.r
\name{scale01}
\alias{scale01}
\title{Scaling data}
\usage{
scale01(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Scale data, making them go from 0 to 1
}
\examples{

x <- seq(-3,3,l = 100)

range(x)

y <- scale01(x)

range(y)

plot(x,y)

}
