% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_auto.R
\name{scale_auto}
\alias{scale_auto}
\title{Auto scale}
\usage{
scale_auto(axis = "x", n = 10, expand = 0.01)
}
\arguments{
\item{axis}{axis indicator: x or y}

\item{n}{number of ticks}

\item{expand}{expanded space}
}
\value{
ggplot2 object
}
\description{
Provide automatic scaling to ggplot2
}
\examples{
library(ggplot2)

plot <-
ggplot(mtcars,aes(hp,drat))+
geom_point()

plot+
 scale_auto("x",n = 10)+
 scale_auto("y",n = 10)


}
