% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_missing_values.r
\name{stat_missing_values}
\alias{stat_missing_values}
\title{Calculate the percentage of missing values}
\usage{
stat_missing_values(df, plot = TRUE)
}
\arguments{
\item{df}{A single data.frame}

\item{plot}{A boolean value, if TRUE the function will plot the percentage of missing values (default = FALSE)}
}
\value{
A tibble.
}
\description{
percentage of missing values for each variable.
}
\examples{


vector <- c(1:10,rep(NA,4))

set.seed(123);data <-
data.frame(
 a = sample(x = vector,size = 100,replace = TRUE),
 b = sample(x = vector,size = 100,replace = TRUE),
 c = sample(x = vector,size = 100,replace = TRUE),
 d = sample(x = vector,size = 100,replace = TRUE),
 e = sample(x = vector,size = 100,replace = TRUE),
 f = sample(x = vector,size = 100,replace = TRUE)
 )

stat_missing_values(data)

}
