% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_normality.r
\name{stat_normality}
\alias{stat_normality}
\title{Test the normality of a numeric variable}
\usage{
stat_normality(x, alpha = 0.05, digits = NULL, print = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{alpha}{A numeric single value (default = 0.05), that is the significance level to compute the variable 'is_normal'}

\item{digits}{A integer single value (default = NULL), that will be the number used to round the values}

\item{print}{A boolean value, if TRUE the function will print the values in the viewer (default = FALSE)}
}
\value{
A tibble (7x4) with each test: statistic, p-value and the condicional check of normality, given the alpha.
}
\description{
Test the normality of a numeric variable.
}
\details{
Apply the following normality tests:

\cr - Anderson-Darling
\cr - Cramer-Von Mises
\cr - Kolmogorov-Smirnov
\cr - Lilliefors
\cr - Pearson chi-square
\cr - Shapiro-Francia
\cr - Shapiro-Wilk
}
\examples{

x <- rnorm(100)

stat_normality(x)
}
