% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_two_cat.r
\name{stat_two_cat}
\alias{stat_two_cat}
\title{Summary table for comparative analysis between two categorical variables}
\usage{
stat_two_cat(df, grp_var, vars, filter = "")
}
\arguments{
\item{df}{A single data.frame}

\item{grp_var}{A character variable from a data.frame}

\item{vars}{One or more variables from a data.frame}

\item{filter}{A character single value (default = ''), that will be the value to filter the variables}
}
\value{
A gt table.
}
\description{
Create a frequency table with chi-square statistic, p-value, cramer's v.
}
\examples{


mtcars \%>\%
 dplyr::mutate(vs = paste0("vs = ",vs)) \%>\%
 stat_two_cat(grp_var = vs,vars = c(am,cyl))
}
