% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_two_num.r
\name{stat_two_num}
\alias{stat_two_num}
\title{Summary table for comparative analysis between two numeric variables}
\usage{
stat_two_num(df, grp_var, num_vars)
}
\arguments{
\item{df}{A single data.frame}

\item{grp_var}{A character variable from a data.frame}

\item{num_vars}{One or more variables from a data.frame}
}
\value{
A gt table.
}
\description{
Create a summary table with t.test: statistic, p-value and IC(95\%).
}
\examples{

n <- 20

set.seed(123);df <-
 data.frame(
   grp_var = sample(paste("group", letters[1:2]),size = n,replace = TRUE),
   num_var1 = rnorm(n),
   num_var2 = abs(rnorm(n))
 )

stat_two_num(df,grp_var,c(num_var1,num_var2))
}
