% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_clean.R
\name{str_clean}
\alias{str_clean}
\title{Remove punctuation and/or accent}
\usage{
str_clean(string, accent = TRUE, punct = TRUE, sub = "")
}
\arguments{
\item{string}{character value}

\item{accent}{logical value to remove accent (default = TRUE)}

\item{punct}{logical value to remove punctuation (default = TRUE)}

\item{sub}{character value to replace punctuation (default = "")}
}
\value{
character value
}
\description{
Remove punctuation and/or accent
}
\examples{

string <- "a..;éâ...íõ"

#remove only punctuation
str_clean(string,accent = FALSE,punct = TRUE)

#remove only accent
str_clean(string,accent = TRUE,punct = FALSE)

#remove both
str_clean(string)
}
