% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_char.r
\name{str_extract_char}
\alias{str_extract_char}
\title{Extract a specific character from a string}
\usage{
str_extract_char(string, char = 1)
}
\arguments{
\item{string}{A character vector}

\item{char}{A integer single value (default = 1), that will be the character position to be extracted}
}
\value{
A character vector.
}
\description{
Extract a specific character from a string.
}
\examples{

str_extract_char(string = "abcdef")

str_extract_char(string = "abcdef", char = 2)


}
