% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_select.r
\name{str_select}
\alias{str_select}
\title{Select parts of a string, given patterns}
\usage{
str_select(string, after = NULL, before = NULL)
}
\arguments{
\item{string}{A character vector}

\item{after}{A character single value (default = NULL), that will be the pattern to select after}

\item{before}{A character single value (default = NULL), that will be the pattern to select before}
}
\value{
A character vector.
}
\description{
Select parts of a string: before, after or between patterns.
}
