% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_text.r
\name{str_to_text}
\alias{str_to_text}
\title{Convert case of a string}
\usage{
str_to_text(string, n_char = 3)
}
\arguments{
\item{string}{characther vector}

\item{n_char}{numeric value}
}
\value{
numeric value
}
\description{
str_to_upper to strings with number of characters lower than "n_char" (default = 3)
}
\examples{

string <- c("aaaaa","bb","ccc","dddd")

str_to_text(string)

}
