% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cat.r
\name{summary_cat}
\alias{summary_cat}
\title{Categoric summary}
\usage{
summary_cat(x)
}
\arguments{
\item{string}{A character vector}
}
\value{
tibble
}
\description{
Compute summary statistics for categoric variable
}
\examples{

x <- c(sample(letters,100,re = T),NA)

x

summary_cat(x)

}
