% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_num.r
\name{summary_num}
\alias{summary_num}
\title{Numeric summary}
\usage{
summary_num(x, minimal = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{minimal}{return minimal version (default = FALSE)}
}
\value{
tibble
}
\description{
Compute summary statistics for numeric variable
}
\examples{

x <- c(rnorm(10),NA,10)

x

summary_num(x)

}
