% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seq.r
\name{summary_seq}
\alias{summary_seq}
\title{Summary sequence of values}
\usage{
summary_seq(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A tibble with the values and the respective number of values repeated in sequence.
}
\description{
Summary of the values in the repeated values in sequence.
}
\examples{

x <- c(1, 1, 1, 2, 2, 2, 2, 3, 4, 5, 6, 7, 1, 1)

summary_seq(x)

}
