% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_chi_square.r
\name{tbl_chi_square}
\alias{tbl_chi_square}
\title{Format data.frame}
\usage{
tbl_chi_square(df, grp_var, vars, filter = "")
}
\arguments{
\item{df}{data.frame}

\item{grp_var}{variable to group by}

\item{vars}{variable(s)}

\item{filter}{value to filter grp_var and vars}
}
\value{
data.frame
}
\description{
Create a frequency table with chi-square statistic, p-value, cramer's v.
}
\examples{

library(dplyr)

}
