% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_compare_num.r
\name{tbl_compare_num}
\alias{tbl_compare_num}
\title{Summary table for comparative analysis}
\usage{
tbl_compare_num(df, grp_var, num_vars, method = "auto")
}
\arguments{
\item{df}{data.frame}

\item{grp_var}{variable to group by}

\item{num_vars}{numerical variable(s)}

\item{method}{method of summary and test (default = 'auto')}
}
\value{
data.frame
}
\description{
Create a summary table with t.test: statistic, p-value and IC(95%).
}
\examples{

set.seed(123);df <-
 data.frame(
   grp_var = sample(paste("group", letters[1:2]),size = 100,replace = TRUE),
   num_var1 = rnorm(100),
   num_var2 = rpois(100,2),
   num_var3 = rexp(100,2),
  num_var4 = abs(rnorm(100))
 )

tbl_compare_num(df,grp_var,c(num_var1,num_var2,num_var3,num_var4),method = c("mean","median","median","mean"))
}
