% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_format_num.r
\name{tbl_format_num}
\alias{tbl_format_num}
\title{Format data.frame}
\usage{
tbl_format_num(df, digits = 2)
}
\arguments{
\item{df}{data.frame}

\item{digits}{number of digits}
}
\value{
data.frame
}
\description{
Apply num_format to all numeric columns in a data.frame
}
\examples{

library(dplyr)

mtcars \%>\%
  count(vs,am) \%>\%
  tbl_format_num(digits = 5)
}
