% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_num_format.R
\name{tbl_num_format}
\alias{tbl_num_format}
\title{Format data.frame}
\usage{
tbl_num_format(df, digits)
}
\arguments{
\item{df}{data.frame}

\item{digits}{number of digits}
}
\value{
data.frame
}
\description{
Apply num_format to all numeric columns in a data.frame
}
\examples{
library(tidyverse)
mtcars \%>\%
  count(vs,am) \%>\%
  tbl_num_format(digits = 5)
}
