% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_plot.r
\name{tbl_plot}
\alias{tbl_plot}
\title{Table plot}
\usage{
tbl_plot(tbl, header_col = "grey75", base_size = 15, bold_last = F, print = T)
}
\arguments{
\item{tbl}{table data.frame}

\item{header_col}{color of the header (default = grey75)}

\item{base_size}{table size (default = 20)}

\item{bold_last}{bold last row (default = F)}
}
\value{
prints a image
}
\description{
Plot a table as an image
}
\examples{

df <- data.frame(grp = c("a","b","c"),
                freq = c(2,4,6))



}
