% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_print.r
\name{tbl_print}
\alias{tbl_print}
\title{Table plot}
\usage{
tbl_print(
  tbl,
  header_color = "grey75",
  text_size = 15,
  bold_last_row = FALSE,
  print = TRUE,
  img_save = FALSE,
  img_path = getwd(),
  img_name = "test",
  img_width = 20,
  img_height = 14
)
}
\arguments{
\item{tbl}{data.frame}

\item{header_color}{color of the header (default = grey75)}

\item{text_size}{table size (default = 20)}

\item{bold_last_row}{bold last row (default = FALSE)}

\item{print}{print the table (default = TRUE)}

\item{img_save}{save the table as image (default = FALSE)}

\item{img_path}{path to save the image (default = working directory)}

\item{img_name}{name of the image to be saved (default = 'test')}

\item{img_width}{width of the image to be saved, in centimeters (default = 20)}

\item{img_height}{height of the image to be saved, in centimeters  (default = 14)}
}
\value{
prints a image
}
\description{
Plot a table as an image
}
\examples{

df <- data.frame(grp = c("a","b","c"),
                 freq = c(2,4,6))

tbl_print(df,bold_last_row = TRUE,header_color = "red")


}
