% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_two_cat.r
\name{tbl_two_cat}
\alias{tbl_two_cat}
\title{Summary table for comparative analysis between two categorical variables}
\usage{
tbl_two_cat(df, grp_var, vars, filter = "")
}
\arguments{
\item{df}{data.frame}

\item{grp_var}{variable to group by}

\item{vars}{variable(s)}

\item{filter}{value to filter grp_var and vars}
}
\value{
data.frame
}
\description{
Create a frequency table with chi-square statistic, p-value, cramer's v.
}
\examples{

library(dplyr)

}
