% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.r
\name{theme_map}
\alias{theme_map}
\title{ggplot2 theme}
\usage{
theme_map(base_size = 14, base_family = "", margin = 0.25)
}
\value{
theme object
}
\description{
no grid, text and ticks
}
\examples{
library(ggplot2)

plot <-
 ggplot(mtcars,aes(hp,drat))+
 geom_point()

plot

plot + theme_map()

}
