% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xy.r
\name{theme_xy}
\alias{theme_xy}
\title{ggplot2 theme}
\usage{
theme_xy(base_size = 14, base_family = "", margin = 0.35)
}
\value{
ggplot2 theme
}
\description{
theme_bw with no minor grid
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + theme_xy()

}
