% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_y.r
\name{theme_y}
\alias{theme_y}
\title{ggplot2 theme}
\usage{
theme_y(base_size = 14, base_family = "")
}
\value{
ggplot2 theme
}
\description{
no minor grid and major x grid
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + theme_y()

}
