% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_digit.r
\name{two_digit}
\alias{two_digit}
\title{Two Digits}
\usage{
two_digit(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector
}
\description{
Format numbers < 10, e.g., 1 as "01"
}
\examples{

x <- c(1,4,10,12)

two_digit(x)

}
