% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataset_documentation.R
\docType{data}
\name{vfx_watermark_white}
\alias{vfx_watermark_white}
\title{Watermark example}
\format{
A three-dimensional array (512x512x4).
}
\usage{
data(vfx_watermark_white)
}
\description{
A object containing a imported .png watermark.
}
\keyword{watermark}
