% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_mark.r
\name{water_mark}
\alias{water_mark}
\title{Water mark}
\usage{
water_mark(png_file)
}
\arguments{
\item{png_file}{png file.}
}
\value{
object to use in ggplot2
}
\description{
Import png file as watermark for ggplot2
}
\examples{


library(ggplot2)

url <- "https://www.osbsoftware.com.br/admin/upload_produtos/RStudio\%20-\%20produto.png"

logo <- water_mark(RCurl::getURLContent(url))

ggplot()+
  annotation_custom(logo)

}
