% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_num.r
\name{as_num}
\alias{as_num}
\title{Numeric vector}
\usage{
as_num(x)
}
\arguments{
\item{x}{character vector}
}
\value{
Numeric vector
}
\description{
Coerce a character vector to numeric: removing thousand separator, replacing comma with dots and applyting as.numeric()
}
\examples{

x <- "10.000,50"

as_num(x)

}
