% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_perc.r
\name{as_perc}
\alias{as_perc}
\title{Scaling}
\usage{
as_perc(x, sum = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{sum}{A boolean value, if TRUE the function will add the values (default = FALSE)}
}
\value{
Numeric vector
}
\description{
Transform numeric to percentage
}
\examples{

x <- c(.4,.6)

as_perc(x)

}
