% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_plot.r
\name{cal_plot}
\alias{cal_plot}
\title{Calendar plot}
\usage{
cal_plot(df, facet = "")
}
\arguments{
\item{df}{data.frame from cal_data function}
}
\value{
ggplot2 plot
}
\description{
Calendar plot, from library ggcal
}
\examples{

library(tidyverse)

# Calendar

date <- seq(as.Date("2017-02-01"), as.Date("2017-07-22"), by="1 day")

fill <- rnorm(length(date))

cal_plot(cal_data(date,fill))

# Calendar + a variable

date <- seq(as.Date("2016-01-01"), as.Date("2016-02-28"),by = "1 day")

tibble(dates = rep(date, 3),
      value = rnorm(3*length(date)),
      grp = rep(c("A","B","C"), each = length(date))) \%>\%
 nest(-grp) \%>\%
 mutate(aux = map(data,~cal_data(.$dates,.$value))) \%>\%
 select(-data) \%>\%
 unnest() \%>\%
 cal_plot(facet = grp)

}
