% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_auc.r
\name{calc_auc}
\alias{calc_auc}
\title{Area under the curve (AUC)}
\usage{
calc_auc(x, y, limits = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{limits}{numeric range for integration (default = x range)}
}
\value{
calc_auc
}
\description{
Calculates AUC
}
\examples{

x <- seq(-3,3,l = 100)

y <- dnorm(x)

plot(x,y)

calc_auc(x,y)

}
