% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_correlation.r
\name{calc_correlation}
\alias{calc_correlation}
\title{Correlation coefficents computation}
\usage{
calc_correlation(x, y, type = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{type}{A character single value, that is the type of the correlation to be computed}
}
\value{
A numeric single value with the computed value.
}
\description{
Computation of correlation coefficients.
}
\details{
Currently, it is possible to compute the following metrics:

\cr - Kendall's Tau correlation coefficient
\cr - Pearson linear correlation coefficient
\cr - Spearmann correlation coefficient
}
\examples{

x <- rnorm(100,1)

calc_correlation(x,x)

}
