% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cramers_v.r
\name{calc_cramers_v}
\alias{calc_cramers_v}
\title{calc_cramers_v}
\usage{
calc_cramers_v(chi_square)
}
\arguments{
\item{chi_square}{object from the application of the function chisq.test}
}
\value{
numeric vector
}
\description{
Computes Cramer's V
}
\examples{

x <- matrix(c(12, 5, 7, 7), ncol = 2)

chi_square <- chisq.test(x)

calc_cramers_v(chi_square)


}
