% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cv.r
\name{calc_cv}
\alias{calc_cv}
\title{Coefficient of variation (CV)}
\usage{
calc_cv(x, as_perc = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{as_perc}{as percentage (default = TRUE)}
}
\value{
Coefficient of variation
}
\description{
Coefficient of variation
}
\examples{

x <- rnorm(100,1)

calc_cv(x)

}
