% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_date_aux.r
\name{calc_date_aux}
\alias{calc_date_aux}
\title{Create auxiliar variables from a date}
\usage{
calc_date_aux(df, dt_var)
}
\arguments{
\item{df}{data.frame}

\item{dt_var}{date variable}
}
\value{
data.frame
}
\description{
Create auxiliar variables from a date, such as: year, month, day, etc.
}
\examples{

calc_date_aux(df,dt)

}
