% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_date_range.r
\name{calc_date_range}
\alias{calc_date_range}
\title{Date range computation}
\usage{
calc_date_range(x)
}
\arguments{
\item{x}{A date vector}
}
\value{
A character single value in the format "{min(x)} - {max(x)}".
}
\description{
The computation of the range of a date vector.
}
\examples{
x <- seq(as.Date("1910/1/1"), as.Date("1911/1/1"), "days")

calc_date_range(x)

}
