% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geometric_mean.r
\name{calc_geometric_mean}
\alias{calc_geometric_mean}
\title{Geometric mean computation}
\usage{
calc_geometric_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric single value with the computed geometric mean.
}
\description{
Computation of the geometric mean.
}
\examples{

x <- rexp(100,.5)

calc_geometric_mean(x)

}
