% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_harmonic_mean.r
\name{calc_harmonic_mean}
\alias{calc_harmonic_mean}
\title{Harmonic mean computation}
\usage{
calc_harmonic_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric single value with the computed harmonic mean.
}
\description{
Computation of the harmonic mean.
}
\examples{

x <- rexp(100,.5)

calc_harmonic_mean(x)

}
