% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.r
\name{calc_mean}
\alias{calc_mean}
\title{Mean computation}
\usage{
calc_mean(x, type = "arithmetic")
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric single value with the computed mean.
}
\description{
Computation of the mean (arithmetic,geometric and harmonic).
}
\examples{

x <- rexp(100,.5)

calc_mean(x,"geometric")

}
