% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.r
\name{calc_mean}
\alias{calc_mean}
\title{Arithmetic, geometric and harmonic mean computation}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A tibble (1x3) with the computed arithmetic, geometric and harmonic mean.
}
\description{
Compute the arithmetic, geometric and harmonic mean.
}
\examples{

x <- rnorm(100)

calc_mean(x)

}
