% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.r
\name{calc_mean}
\alias{calc_mean}
\title{Mean computation}
\usage{
calc_mean(
  x,
  type = c("arithmetic", "geometric", "harmonic"),
  trim = NULL,
  weight = NULL
)
}
\arguments{
\item{x}{A numeric vector}

\item{type}{A character vector}

\item{trim}{A numeric vector, that will be fraction of data to be trimmed from both ends}

\item{weight}{A numeric vector, that will be the weighted of the arithmetic mean}
}
\value{
A numeric single value with the computed mean.
}
\description{
Computation of the mean (arithmetic,geometric and harmonic).
}
\examples{

x <- rexp(100,.5)

calc_mean(x,"geometric")

}
