% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.r
\name{calc_mode}
\alias{calc_mode}
\title{Mode computation}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A vector with the computed mode(s).
}
\description{
Computation of the mode.
}
\examples{

table(mtcars$cyl)

calc_mode(mtcars$cyl)

}
