% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_peak_density.r
\name{calc_peak_density}
\alias{calc_peak_density}
\title{Peak density value computation}
\usage{
calc_peak_density(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric single value with the computed peak density value.
}
\description{
Computation of the peak value, of a density function with a gaussian kernel.
}
\examples{

x <- rnorm(100)

plot(density(x))

calc_peak_density(x)

}
