% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{expand_grid without repetition}
\usage{
expand_grid_unique(x, y, include.equals = FALSE)
}
\arguments{
\item{y}{}
}
\value{

}
\description{

}
\examples{


}
