% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{All combinations of inputs without repetitions}
\usage{
expand_grid_unique(x, y, include_equals = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector}

\item{include_equals}{A boolean value, if TRUE the function will ignore the equal pairs (default = FALSE)}
}
\value{
A tibble with all the combinations of x and y.
}
\description{
Create a data frame from all combinations of the supplied vectors,
 without considering repetitions.
}
\examples{

expand_grid_unique(x = 1:6,y = 1:6)

}
