% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_digit.r
\name{format_digit}
\alias{format_digit}
\title{two_digit}
\usage{
format_digit(x, digits = 2)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
character vector
}
\description{
Format numbers < 10, e.g., 1 as "01"
}
\examples{

x <- c(1,4,10,12)

format_digit(x)

}
