% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.r
\name{format_num}
\alias{format_num}
\title{Format numeric data}
\usage{
format_num(
  x,
  digits = 2,
  decimal_mark = ".",
  thousand_mark = ",",
  br_mark = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{decimal_mark}{A character single value to use as the decimal mark}

\item{thousand_mark}{A character single value to use as the thousand mark}

\item{br_mark}{A boolean value, if TRUE the function will set decimal_mark = ',' and thousand_mark = '.' (default = FALSE)}
}
\value{
formatted data
}
\description{
Format data as "10.500,25"
}
\examples{

format_num(12345.67)

}
