% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.r
\name{format_num}
\alias{format_num}
\title{Format numeric value to add marks}
\usage{
format_num(
  x,
  digits = 2,
  decimal_mark = ".",
  thousand_mark = ",",
  br_mark = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{decimal_mark}{A character single value to use as the decimal mark (default = '.')}

\item{thousand_mark}{A character single value to use as the thousand mark (default = ',')}

\item{br_mark}{A boolean value, if TRUE the function will set decimal_mark = ',' and thousand_mark = '.' (default = FALSE)}
}
\value{
A character vector with the formatted vector.
}
\description{
Format value to add a thousand and decimal mark.
}
\examples{

format_num(12345.67)

}
