% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_value.r
\name{format_p_value}
\alias{format_p_value}
\title{Format p value}
\usage{
format_p_value(p_value, lower_bound = NULL, ...)
}
\arguments{
\item{p_value}{A numeric vector}

\item{lower_bound}{A numeric single value (default = NULL), that will be the lower bound}
}
\value{
Characther vector
}
\description{
Format p value
}
\examples{

format_p_value(c(.001,.00000001),lower_bound = 0.001)

}
