% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_scale.r
\name{format_scale}
\alias{format_scale}
\title{Scaling data}
\usage{
format_scale(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{new_min}{A numeric single value (default = 0), that will be the new minimum value}

\item{new_max}{A numeric single value (default = 1), that will be the new maximum value}
}
\value{
numeric vector
}
\description{
Rescale data
}
\examples{

x <- seq(-3,3,l = 100)

range(x)

y <- format_scale(x)

range(y)

plot(x,y)

}
