% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_datetime.r
\name{is_datetime}
\alias{is_datetime}
\title{Conditional check for a datetime variable}
\usage{
is_datetime(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the vector is a datetime.
}
