% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.r
\name{is_even}
\alias{is_even}
\title{Conditional check for an even value}
\usage{
is_even(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is even.
}
\examples{

is_even(4)

is_even(3)

}
