% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.r
\name{is_integer}
\alias{is_integer}
\title{Check if a value is a integer, but not by their type}
\usage{
is_integer(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Numeric vector
}
\description{
return TRUE if the number is a integer, not considering their type but the number itself
}
\examples{

is_integer(1)

is_integer(1.1)

}
