% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_negative.r
\name{is_negative}
\alias{is_negative}
\title{Negative number}
\usage{
is_negative(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector
}
\description{
return TRUE if the number is negative
}
\examples{

is_negative(4)

}
