% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.r
\name{is_outlier}
\alias{is_outlier}
\title{Conditional check for an outlier value}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is an outlier (boxplot criteria).
}
\examples{

x <- c(rnorm(10),10)

x

is_outlier(x)

}
