% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_positive.r
\name{is_positive}
\alias{is_positive}
\title{Positive number}
\usage{
is_positive(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Numeric vector
}
\description{
return TRUE if the number is positive
}
\examples{

is_positive(4)

}
