% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_string.r
\name{is_string}
\alias{is_string}
\title{String value}
\usage{
is_string(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
Boolean vector
}
\description{
return TRUE if the variable is a string (character or factor)
}
\examples{

is_string("A")

}
