% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnot_in.r
\name{isnot_in}
\alias{isnot_in}
\title{Opposite of in}
\usage{
isnot_in(x, y)
}
\arguments{
\item{x}{A  single value}

\item{y}{A  vector}
}
\value{
boolean
}
\description{
Check if  x is not contained in y
}
\examples{

isnot_in("a", letters)

}
