% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnot_na.r
\name{isnot_na}
\alias{isnot_na}
\title{Conditional check if a value is not a NA}
\usage{
isnot_na(x)
}
\arguments{
\item{x}{A  vector}
}
\value{
A boolean vector with the return from the condition check.
}
\description{
Return TRUE if the value is not a NA.
}
\examples{

isnot_na(2)

isnot_na(NA)

}
