% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_summary.r
\name{num_summary}
\alias{num_summary}
\title{Numeric summary}
\usage{
num_summary(x, y)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Numeric vector
}
\description{
Compute summary statistics for numeric variable
}
\examples{

x <- rnorm(100)

num_summary(x)

}
