% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_qua.R
\name{palette_qua}
\alias{palette_qua}
\title{Palettes for qualitative scale}
\usage{
palette_qua(
  name = c("bojack_horseman", "cowboy_bebop", "doctor_who", "final_space",
    "gurren_lagann", "heartstopper", "homeland", "kick_ass", "legion", "omitb",
    "moonlight", "ms_marvel", "pulp_fiction", "red_dead", "rick_and_morty", "star_trek",
    "this_is_us", "versace", "ww84"),
  reverse = FALSE
)
}
\arguments{
\item{name}{A character vector, that is the pallete name}

\item{reverse}{A boolean value, if TRUE the function will reverse the pallete order (default = FALSE)}
}
\value{
A character vector with the colors hex codes.
}
\description{
Palettes for a qualitative scale, with the following palettes:
 \cr
 \cr - bojack_horseman: 5 colors, inspired by the animation BoJack Horseman.
 \cr - cowboy_bebop: 5 colors, inspired by the animation Cowboy Bebop.
 \cr - doctor_who: 5 colors, inspired by the TV series Doctor Who.
 \cr - final_space: 4 colors, inspired by the animation Final Space.
 \cr - gurren_lagann: 5 colors, inspired by the animation Tengen Toppa Gurren-Lagann.
 \cr - heartstopper: 5 colors, inspired by the TV series Heartstopper.
 \cr - homeland: 5 colors, inspired by the TV series Homeland.
 \cr - kick_ass: 6 colors, inspired by the movie Kick-ass.
 \cr - legion: 9 colors, inspired by the TV series Legion.
 \cr - moonlight: 6 colors, inspired by the movie Moonlight.
 \cr - ms_marvel: 6 colors, inspired by the TV Series Ms Marvel.
 \cr - omitb: 5 colors, inspired by the TV series Only Murders in the Building.
 \cr - pulp_fiction: 8 colors, inspired by the movie Pulp Fiction.
 \cr - red_dead: 6 colors, inspired by the video game Red Dead Redemption.
 \cr - rick_and_morty: 8 colors, inspired by the animation Rick and Morty.
 \cr - star_trek: 5 colors, inspired by the TV series Star Trek.
 \cr - this_is_us: 5 colors, inspired by the TV series This is Us.
 \cr - versace: 8 colors, inspired by the TV series Versace.
 \cr - ww84: 6 colors, inspired by the movie Wonder Woman 84.
}
\examples{

palette_example(palette = palette_qua(name = "this_is_us"))

}
