% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_seq.R
\name{palette_seq}
\alias{palette_seq}
\title{Palettes for sequential scale}
\usage{
palette_seq(
  name = c("andor", "badlands", "breaking_bad", "clockwork_orange", "cyberpunk",
    "fleabag", "house_md", "house_of_cards", "loki", "lovecraft_country",
    "mare_easttown", "peaky_blinders", "raised_by_wolves", "sabrina", "sandman",
    "sucession", "the_wire", "tiger_king", "westworld1", "westworld2"),
  reverse = FALSE
)
}
\arguments{
\item{name}{A character vector, that is the pallete name}

\item{reverse}{A boolean value, if TRUE the function will reverse the pallete order (default = FALSE)}
}
\value{
A character vector with the colors hex codes.
}
\description{
Palettes for a sequential scale, with the following palettes:
 \cr
 \cr - andor: 6 colors, inspired by the TV series Andor (Star Wars).
 \cr - badlands: 5 colors, inspired by the TV series Into the Badlands.
 \cr - breaking_bad: 5 colors, inspired by the TV series Breaking Bad.
 \cr - clockwork_orange: 8 colors, inspired by the movie Clockwork Orange.
 \cr - cyberpunk: 5 colors, inspired by the videogame Cyberpunk 2077.
 \cr - fleabag: 6 colors, inspired by the TV series Fleabag.
 \cr - house_md: 6 colors, inspired by the TV series House MD.
 \cr - house_of_cards: 7 colors, inspired by the TV series House of Cards.
 \cr - loki: 7 colors, inspired by the TV series Loki.
 \cr - lovecraft_country: 9 colors, inspired by the TV series Lovecraft Country.
 \cr - mare_easttown: 7 colors, inspired by the TV series Mare of EastTown.
 \cr - peaky_blinders: 8 colors, inspired by the TV Peaky Blinders.
 \cr - raised_by_wolves: 6 colors, inspired by the TV series Raised by Wolves.
 \cr - sabrina: 8 colors, inspired by the TV series Sabrina.
 \cr - sandman: 6 colors, inspired by the TV series Sandman.
 \cr - sucession: 6 colors, inspired by the TV series Sucession.
 \cr - the_wire: 6 colors, inspired by the TV series The Wire.
 \cr - tiger_king: 7 colors, inspired by the TV series Tiger King.
 \cr - westworld1: 7 colors, inspired by the TV series WestWorld.
 \cr - westworld2: 8 colors, inspired by the TV series WestWorld.
}
\examples{

palette_example(palette = palette_seq(name = "cyberpunk"))

}
