% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_no_background.r
\docType{data}
\name{plt_no_background}
\alias{plt_no_background}
\title{ggplot2: Apply a transparent background.}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
plt_no_background
}
\value{
A ggplot2 object.
}
\description{
Apply a transparent background for a ggplot2 plot.
}
\examples{
library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_no_background

}
\keyword{datasets}
