% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_scale_auto.r
\name{plt_scale_auto}
\alias{plt_scale_auto}
\title{Auto scale}
\usage{
plt_scale_auto(
  axis = "x",
  n_breaks = 10,
  expand = 0.01,
  thousand_mark = ".",
  decimal_mark = ","
)
}
\arguments{
\item{axis}{A character single value (default = 'x'), that will be the axis to scale, must be x, y or xy}

\item{n_breaks}{A numeric integer single value (default = 10), that will be the number of axis breaks}

\item{expand}{A numeric single value (default = 0.01), that will be the expanded margin to the axis}

\item{thousand_mark}{A character single value to use as the thousand mark}

\item{decimal_mark}{A character single value to use as the decimal mark}
}
\value{
ggplot2 object
}
\description{
Provide automatic scaling to ggplot2
}
\examples{
library(ggplot2)

plot <-
ggplot(mtcars,aes(hp,drat))+
geom_point()

plot+
 plt_scale_auto("x",n_breaks = 10)+
 plt_scale_auto("y",n_breaks = 10)


}
