% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_theme_x.r
\name{plt_theme_x}
\alias{plt_theme_x}
\title{ggplot2: A black and white theme without major grid lines in the y axis}
\usage{
plt_theme_x(base_size = 14, base_family = "", margin = 0.35)
}
\arguments{
\item{base_size}{A numeric single value (default = 14), that will be the base font size}

\item{margin}{A numeric single value (default = 0.35), that will be the plot margin}
}
\value{
A ggplot2 object.
}
\description{
A theme with no minor grids in both axes and no major grid lines in the y axis.
}
\examples{

library(ggplot2)

plot <-
  ggplot(mtcars,aes(hp,drat))+
  geom_point()

plot

plot + plt_theme_x()

}
