% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_water_mark.r
\name{plt_water_mark}
\alias{plt_water_mark}
\title{Water mark}
\usage{
plt_water_mark(png_file, local_file = TRUE)
}
\arguments{
\item{png_file}{png file path}

\item{local_file}{if is a local file, local = T (default), if it is a url local = F}
}
\value{
object to use in ggplot2
}
\description{
Import png file as watermark for ggplot2
}
\examples{


library(ggplot2)

url <- "https://upload.wikimedia.org/wikipedia/commons/thumb/1/1b/R_logo.svg/1200px-R_logo.svg.png"

logo <- plt_water_mark(url,local_file = F)

ggplot()+
  annotation_custom(logo)
}
