% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_water_mark.r
\name{plt_water_mark}
\alias{plt_water_mark}
\title{ggplot2: Application of a watermark}
\usage{
plt_water_mark(obj)
}
\arguments{
\item{obj}{The name of the R object with the watermark as an array}
}
\value{
A ggplot2 object.
}
\description{
Application of a watermark for a ggplot2 plot.
}
\examples{

library(ggplot2)

ggplot(mtcars)+
plt_water_mark(vfx_watermark)

}
