% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_water_mark.r
\name{plt_water_mark}
\alias{plt_water_mark}
\title{Water mark}
\usage{
plt_water_mark(png_file)
}
\arguments{
\item{png_file}{png file path}
}
\value{
object to use in ggplot2
}
\description{
Import png file as watermark for ggplot2
}
\examples{


library(ggplot2)

#A local png file is necessary
logo <- plt_water_mark(path_to_local_png)

ggplot()+
  annotation_custom(logo)
}
