% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pearson.r
\name{r_pearson}
\alias{r_pearson}
\title{Correlated linear data generator}
\usage{
r_pearson(n, p = 0.5, ...)
}
\arguments{
\item{n}{number of observations}

\item{p}{pearson linear correlation coefficient}
}
\value{
data.frame with 2 numeric correlated
}
\description{
Simulates
}
\examples{

plot(r_pearson(100,.8,mean = 3))

}
