% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_number_unique.r
\name{row_number_unique}
\alias{row_number_unique}
\title{row_number() but consdering unique values}
\usage{
row_number_unique(x)
}
\arguments{
\item{x}{vector}
}
\value{
Numeric vector
}
\description{
Get the row number for unique values
}
\examples{

row_number_unique(mtcars$gear)

}
