% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpearson.r
\name{rpearson}
\alias{rpearson}
\title{Linear correlated data generator}
\usage{
rpearson(n = 25, pearson = 0.5, tol = 0.1, mean = 0, sd = 1)
}
\arguments{
\item{n}{Number of observations to simulate}

\item{pearson}{Pearson linear correlation coefficient}

\item{tol}{Tolerance between the simulated and estimated correlation}

\item{mean}{Mean of the simulated data}

\item{sd}{Standard deviation of the simulated data}
}
\value{
data.frame with 2 numeric correlated series and both the simulated and estimated pearson linear coefficient
}
\description{
Simulates linear correlated data
}
\examples{

df <- rpearson(n = 100, pearson = .8, mean = 3)

head(df)

plot(df)

}
