% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_clean.r
\name{str_clean}
\alias{str_clean}
\title{Remove punctuation and/or accent}
\usage{
str_clean(string, remove_accent = TRUE, remove_punct = TRUE, sub_punct = "")
}
\arguments{
\item{string}{character value}

\item{remove_accent}{logical value to remove accent (default = TRUE)}

\item{remove_punct}{logical value to remove punctuation (default = TRUE)}

\item{sub_punct}{character value to replace punctuation (default = "")}
}
\value{
character value
}
\description{
Remove punctuation and/or accent
}
\examples{

string <- "a..;éâ...íõ"

#remove only punctuation
str_clean(string,remove_accent = FALSE,remove_punct = TRUE)

#remove only accent
str_clean(string,remove_accent = TRUE,remove_punct = FALSE)

#remove both
str_clean(string)
}
