% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_clean.R
\name{str_clean}
\alias{str_clean}
\title{Remove punctuation and/or accent from a string}
\usage{
str_clean(string, remove_accent = TRUE, remove_punct = TRUE, sub_punct = "")
}
\arguments{
\item{string}{A character vector}

\item{remove_accent}{A boolean value, if TRUE the function will remove accent symbols from the string (default = TRUE)}

\item{remove_punct}{A boolean value, if TRUE the function will remove punctuation symbols from the string (default = TRUE)}

\item{sub_punct}{A character single value (default = ''), that will be replacement for the punctuation symbols}
}
\value{
A character vector.
}
\description{
Remove punctuation and/or accent from a string.
}
\examples{

string <- "a..;éâ...íõ"

#remove only punctuation
str_clean(string,remove_accent = FALSE,remove_punct = TRUE)

#remove only accent
str_clean(string,remove_accent = TRUE,remove_punct = FALSE)

#remove both
str_clean(string)
}
