% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_keep.r
\name{str_keep}
\alias{str_keep}
\title{Keep only a type of character in the string}
\usage{
str_keep(string, keep = "text")
}
\arguments{
\item{string}{A character vector}

\item{keep}{A character single value (default = 'text'), that will be the type of character to keep}
}
\value{
Character vector
}
\description{
Keep only a type of character in the string
}
\examples{

string <- "1ABCF45Z89$$\%#"

str_keep(string,keep = "text")

str_keep(string,keep = "numbers")

str_keep(string,keep = "special")
}
