% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_select.R
\name{str_select}
\alias{str_select}
\title{Select string}
\usage{
str_select(string, after = NULL, before = NULL)
}
\arguments{
\item{string}{character value}

\item{after}{first pattern}

\item{before}{second pattern}
}
\value{
character value
}
\description{
Select string: before, after or between patterns
}
\examples{

string <- "example text STRING1 TARGET STRING2 example text"

#Select a string, before a pattern
str_select(string,before = "STRING2")

#Select a string, after a pattern
str_select(string,after = "STRING1")

#Select a string, between two patterns
str_select(string,"STRING1","STRING2")
}
