% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_select.r
\name{str_select}
\alias{str_select}
\title{Select parts of a string, given patterns}
\usage{
str_select(string, after = NULL, before = NULL)
}
\arguments{
\item{string}{A character vector}

\item{after}{A character single value (default = NULL), that will be the pattern to select after}

\item{before}{A character single value (default = NULL), that will be the pattern to select before}
}
\value{
A character vector.
}
\description{
Select parts of a string: before, after or between patterns.
}
\examples{

string <- "begin STRING1 TARGET STRING2 end"
#Select a string, before a pattern
str_select(string,before = "STRING2")

#Select a string, after a pattern
str_select(string,after = "STRING1")

#Select a string, between two patterns
str_select(string,"STRING1","STRING2")
}
