% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_abb.r
\name{str_to_abb}
\alias{str_to_abb}
\title{Convert case of a string}
\usage{
str_to_abb(string, n_char = 3)
}
\arguments{
\item{string}{A character vector}

\item{n_char}{A numeric integer single value (default = 3), that will be the number of characters to define as a abbreviation}
}
\value{
A character vector.
}
\description{
Apply str_to_upper to strings with number of characters lower than "n_char".
}
\examples{

string <- c("aaaaa","bb","ccc","dddd")

str_to_abb(string)

}
