% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_text.R
\name{str_to_text}
\alias{str_to_text}
\title{Convert case of a string}
\usage{
str_to_text(x, n_char = 3)
}
\arguments{
\item{x}{characther vector}

\item{n_char}{numeric value}
}
\value{
numeric value
}
\description{
str_to_upper to strings with number of characters lower than "n_char" (default = 3)
}
\examples{

x <- rnorm(100)

num_mode(x)

}
