% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data.r
\name{summary_data}
\alias{summary_data}
\title{Data summary}
\usage{
summary_data(df)
}
\arguments{
\item{df}{A single data.frame}
}
\value{
tibble
}
\description{
Compute summary statistics for data
}
\examples{

mtcars

summary_data(mtcars)

}
