% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data.r
\name{summary_data}
\alias{summary_data}
\title{Summary data.frame}
\usage{
summary_data(df)
}
\arguments{
\item{df}{A single data.frame}
}
\value{
A diagnostic of the data.frame.
}
\description{
Summary statistics for a data.frame.
}
\examples{

mtcars

summary_data(mtcars)

}
