% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_num.r
\name{summary_num}
\alias{summary_num}
\title{Summary  of numeric variables}
\usage{
summary_num(x, minimal = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{minimal}{A boolean value, if TRUE the function will return a minimal version of the summary (default = FALSE)}
}
\value{
A tibble with:
\cr
\cr - n: the number of observations;
\cr - na: the number of missing values;
\cr - negative: the number of negative values;
\cr - equal_zero: the number of values equal to zero;
\cr - positive: the number of positive values;
\cr - min: the minimum;
\cr - p25: the first quartile;
\cr - p50: the second quartile (median);
\cr - p75: the third quartile;
\cr - max: the maximum;
\cr - mode: the peak density value;
\cr - mean: the mean;
\cr - cv: the coefficient of variation.
}
\description{
Summary statistics for numeric variables.
}
\examples{

x <- c(rnorm(10),NA,10)

x

summary_num(x)

}
