% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_num.r
\name{summary_num}
\alias{summary_num}
\title{Summary of a numeric variable}
\usage{
summary_num(
  x,
  type = FALSE,
  other_means = FALSE,
  skewness = FALSE,
  kurtosis = FALSE
)
}
\arguments{
\item{x}{A numeric vector}

\item{type}{A boolean value, if TRUE the function will add metrics related to the variables type (default = TRUE)}

\item{other_means}{A boolean value, if TRUE the function will add the harmonic and geometric means (default = FALSE)}

\item{skewness}{A boolean value, if TRUE the function will add the skewness coefficients (default = FALSE)}

\item{kurtosis}{A boolean value, if TRUE the function will add the kurtosis coefficients (default = FALSE)}
}
\value{
A tibble with the summary metrics.
}
\description{
Summary statistics for numeric variables.
}
\details{
By default the summary statistics are:
\cr - min: the minimum;
\cr - p25: the first quartile;
\cr - p50: the second quartile (median);
\cr - p75: the third quartile;
\cr - max: the maximum;
\cr - mode: the peak density value;
\cr - mean: the mean;
\cr - cv: the coefficient of variation.

If `type` = TRUE, the following metrics will be added:
\cr - n: the number of observations;
\cr - na: the number of missing values;
\cr - negative: the number of negative values;
\cr - equal_zero: the number of values equal to zero;
\cr - positive: the number of positive values.

If `other_means` = TRUE, the following metrics will be added:
\cr - geometric_mean: the geometric mean;
\cr - harmonic_mean: the harmonic mean.

If `skewness` = TRUE, the following metrics will be added:
\cr - Bowley
\cr - Fisher-Pearson
\cr - Kelly
\cr - Rao
\cr - Pearson median

If `kurtosis` = TRUE, the following metrics will be added:
\cr - Bowley
\cr - Fisher-Pearson
\cr - Kelly
\cr - Rao
\cr - Pearson median
}
\examples{

x <- c(rnorm(10),NA,10)

x

summary_num(x)

}
