% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seq.r
\name{summary_seq}
\alias{summary_seq}
\title{Compute the number of values that are repeated in sequence}
\usage{
summary_seq(x)
}
\arguments{
\item{x}{vector}
}
\value{
data.frame
}
\description{
Compute the number of values that are repeated in sequence
}
\examples{

x <- c(1, 1, 1, 2, 2, 2, 2, 3, 4, 5, 6, 7, 1, 1)

summary_seq(x)

}
