% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_print.r
\name{tbl_print}
\alias{tbl_print}
\title{Table plot}
\usage{
tbl_print(
  tbl,
  header_col = "grey75",
  base_size = 15,
  bold_last = FALSE,
  print = TRUE,
  save = FALSE,
  img_path = getwd(),
  img_name = "test",
  img_width = 20,
  img_height = 14
)
}
\arguments{
\item{tbl}{table data.frame}

\item{header_col}{color of the header (default = grey75)}

\item{base_size}{table size (default = 20)}

\item{bold_last}{bold last row (default = F)}
}
\value{
prints a image
}
\description{
Plot a table as an image
}
\examples{

df <- data.frame(grp = c("a","b","c"),
                freq = c(2,4,6))

tbl_print(df,bold_last = T,header_col = "red")


}
