% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_mark.r
\name{water_mark}
\alias{water_mark}
\title{Water mark}
\usage{
water_mark(png_file)
}
\arguments{
\item{png_file}{png file.}
}
\value{
object to use in ggplot2
}
\description{
Import png file as watermark for ggplot2
}
